Remove pint and related logic
Remove the uses of pint for the values attribute in RegularTimseries.
The uses were limited and knowledge of the library is much more limited
than numpy, which it wraps.

Future work will be done to add functions to help with common unit
conversions, without changing the types used by the RegularTimseries
object.