Improve string-y equality error handling
Previously, when testing `__eq__()` with a DatasetPath and a string, the
string was converted to a DatasetPath via the classes `from_str`
classmethod. This classmethod can raise an error is the string isn't
valid. This exception went un-caught. This change catches the exception
and returns False for that `__eq__()` call.